pyproject.toml
[tool.poetry]
name = "pgcopy-standalone"
version = "0.1.0"
description = "Fast db insert with postgresql binary copy"
authors = ["Aryeh Leib Taurog <python@aryehleib.com>", "Fang-Pen Lin <hello@fangpenlin.com>"]
license = "MIT"
readme = "README.rst"
packages = [{include = "pgcopy"}]

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2023.3.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
